#51 - перенос Code Generation
