+++ ExtendedJavaScriptSubset.sln
@@ -1,7 +1,5 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HydraScript.Lib", "src\HydraScript.Lib\HydraScript.Lib.csproj", "{83524079-4A56-4AF0-9011-F7CA871536E9}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HydraScript", "src\HydraScript\HydraScript.csproj", "{5DA79C0C-2B98-4E64-81EA-92AFFC5204D3}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HydraScript.Tests", "tests\HydraScript.Tests\HydraScript.Tests.csproj", "{F0DD56CB-C68A-4CDC-AA54-0A07342FE08B}"
@@ -94,10 +92,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{83524079-4A56-4AF0-9011-F7CA871536E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{83524079-4A56-4AF0-9011-F7CA871536E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{83524079-4A56-4AF0-9011-F7CA871536E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{83524079-4A56-4AF0-9011-F7CA871536E9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5DA79C0C-2B98-4E64-81EA-92AFFC5204D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5DA79C0C-2B98-4E64-81EA-92AFFC5204D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5DA79C0C-2B98-4E64-81EA-92AFFC5204D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -136,7 +130,6 @@ Global
 		{C7773DC6-9052-4F69-B947-CBFD5D663E80} = {54CBE5A7-3C3E-44ED-B877-7B08A818083B}
 		{0E31199C-4893-4920-A356-D1FF7850BFAD} = {54CBE5A7-3C3E-44ED-B877-7B08A818083B}
 		{5DA79C0C-2B98-4E64-81EA-92AFFC5204D3} = {FB8F6EE1-1942-46D6-954E-9A1647BBDF10}
-		{83524079-4A56-4AF0-9011-F7CA871536E9} = {FB8F6EE1-1942-46D6-954E-9A1647BBDF10}
 		{F0DD56CB-C68A-4CDC-AA54-0A07342FE08B} = {3F131901-A9EC-451A-B7E9-726887CFE5FB}
 		{A280C99E-0112-4F3C-A5E9-B793F2D9898E} = {FB8F6EE1-1942-46D6-954E-9A1647BBDF10}
 		{F0FC6D23-F932-4C11-8D0F-07894FC61511} = {FB8F6EE1-1942-46D6-954E-9A1647BBDF10}
