+++ src/Application/HydraScript.Application.CodeGeneration/GlobalUsings.cs
@@ -1,0 +1,3 @@
+// Global using directives
+
+global using Visitor.NET;
