+++ src/Application/HydraScript.Application.CodeGeneration/Services/IValueDtoConverter.cs
@@ -1,0 +1,9 @@
+using HydraScript.Domain.BackEnd;
+using HydraScript.Domain.FrontEnd.Parser.Impl.Ast.Nodes.Expressions.PrimaryExpressions;
+
+namespace HydraScript.Application.CodeGeneration.Services;
+
+public interface IValueDtoConverter
+{
+    IValue Convert(ValueDto dto);
+}
