+++ src/Application/HydraScript.Application.StaticAnalysis/Services/IStandardLibraryProvider.cs
@@ -1,0 +1,8 @@
+using HydraScript.Domain.IR;
+
+namespace HydraScript.Application.StaticAnalysis.Services;
+
+public interface IStandardLibraryProvider
+{
+    ISymbolTable GetStandardLibrary();
+}
