+++ src/Application/HydraScript.Application.StaticAnalysis/Services/ITypeDeclarationsResolver.cs
@@ -1,0 +1,10 @@
+using HydraScript.Domain.FrontEnd.Parser.Impl.Ast.Nodes.Declarations;
+
+namespace HydraScript.Application.StaticAnalysis.Services;
+
+public interface ITypeDeclarationsResolver
+{
+    void Store(TypeDeclaration declaration);
+
+    void Resolve();
+}
