--- src/HydraScript.Lib/GlobalUsings.cs
@@ -1,5 +1,0 @@
-// Global using directives
-
-global using Expression = HydraScript.Domain.FrontEnd.Parser.Impl.Ast.Nodes.Expressions.Expression;
-global using Type = HydraScript.Domain.IR.Types.Type;
-global using Visitor.NET;
