--- src/HydraScript.Lib/HydraScript.Lib.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <TreatWarningsAsErrors>True</TreatWarningsAsErrors>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Visitor.NET" Version="4.1.2" />
-      <PackageReference Include="Visitor.NET.AutoVisitableGen" Version="1.3.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <InternalsVisibleTo Include="HydraScript.Tests" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="IR\CheckSemantics\Visitors\" />
-    </ItemGroup>
-
-</Project>
