--- src/HydraScript.Lib/IR/Ast/Visitors/Services/IValueDtoConverter.cs
@@ -1,9 +1,0 @@
-using HydraScript.Lib.BackEnd;
-using HydraScript.Lib.IR.Ast.Impl.Nodes.Expressions.PrimaryExpressions;
-
-namespace HydraScript.Lib.IR.Ast.Visitors.Services;
-
-public interface IValueDtoConverter
-{
-    IValue Convert(ValueDto dto);
-}
