--- src/HydraScript.Lib/IR/CheckSemantics/Visitors/Services/IStandardLibraryProvider.cs
@@ -1,8 +1,0 @@
-using HydraScript.Lib.IR.CheckSemantics.Variables;
-
-namespace HydraScript.Lib.IR.CheckSemantics.Visitors.Services;
-
-public interface IStandardLibraryProvider
-{
-    ISymbolTable GetStandardLibrary();
-}
