--- src/HydraScript.Lib/IR/CheckSemantics/Visitors/Services/ITypeDeclarationsResolver.cs
@@ -1,10 +1,0 @@
-using HydraScript.Lib.IR.Ast.Impl.Nodes.Declarations;
-
-namespace HydraScript.Lib.IR.CheckSemantics.Visitors.Services;
-
-public interface ITypeDeclarationsResolver
-{
-    void Store(TypeDeclaration declaration);
-
-    void Resolve();
-}
