+++ src/HydraScript/HydraScript.csproj
@@ -10,7 +10,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\HydraScript.Lib\HydraScript.Lib.csproj" />
       <ProjectReference Include="..\Infrastructure\HydraScript.Infrastructure\HydraScript.Infrastructure.csproj" />
     </ItemGroup>
